Add init file to code root
